requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==24.2.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.8.2
blockbuster==1.5.24
CacheControl==0.14.2
cachetools==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cloudevents==1.11.0
cloudpickle==3.1.1
comm==0.2.2
cryptography==43.0.3
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docx2txt==0.9
executing==2.2.0
fastapi==0.115.11
fastjsonschema==2.21.1
firebase-admin==6.7.0
firestore==0.0.8
Flask==3.0.3
forbiddenfruit==0.1.4
fqdn==1.5.1
frozenlist==1.5.0
functions-framework==3.5.0
google-api-core==2.22.0
google-api-python-client==2.166.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.3
google-cloud-error-reporting==1.9.1
google-cloud-firestore==2.20.1
google-cloud-logging==3.11.3
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
iso8601==2.1.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.6.1
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jsonschema_rs==0.29.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.54
langchain-mongodb==0.2.0
langchain-openai==0.2.4
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.7
langgraph==0.3.31
langgraph-api==0.1.9
langgraph-checkpoint==2.0.24
langgraph-checkpoint-mongodb==0.1.3
langgraph-cli==0.2.5
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.0.4
langgraph-sdk==0.1.61
langsmith==0.3.32
MarkupSafe==2.1.3
marshmallow==3.26.1
matplotlib-inline==0.1.7
mistune==3.1.3
mongoengine==0.29.1
motor==3.7.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.26.4
openai==1.53.0
opentelemetry-api==1.27.0
orjson==3.10.16
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pusher==3.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
pyinstrument==5.0.1
PyJWT==2.10.1
pymongo==4.11.3
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.2.3
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
rsa==4.9
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.46.1
structlog==25.2.0
tenacity==9.1.2
terminado==0.18.1
tiktoken==0.8.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.66.6
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.0
validators==0.34.0
watchdog==5.0.3
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.0
zipp==3.20.2
zstandard==0.23.0
